% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ch_flows}
\alias{ch_flows}
\title{Swiss Flows Dataset}
\format{
\subsection{\code{ch_flows}}{

A data frame with 676 rows and 3 columns:
\describe{
\item{origin}{A \code{character} vector representing the origin location identifier (must match the \code{id} in the \code{ch_locations} dataset).}
\item{dest}{A \code{character} vector representing the destination location identifier (must match the \code{id} in the \code{ch_locations} dataset).}
\item{count}{An \code{integer} vector representing the flow count between the origin and destination locations.}
}
}
}
\usage{
ch_flows
}
\description{
A dataset containing flow data between various locations in Switzerland. This data represents the flow counts between origin and destination locations, identified by their unique codes.
}
\keyword{datasets}
