% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ch_locations}
\alias{ch_locations}
\title{Swiss Locations Dataset}
\format{
\subsection{\code{ch_locations}}{

A data frame with 26 rows and 4 columns:
\describe{
\item{id}{A \code{character} vector representing the unique identifier for each Swiss location (e.g., "JU", "LU").}
\item{name}{A \code{character} vector representing the name of each location (e.g., "Jura", "Luzern").}
\item{lat}{A \code{numeric} vector representing the latitude of each location in WGS84 (EPSG: 4326) coordinate reference system.}
\item{lon}{A \code{numeric} vector representing the longitude of each location in WGS84 (EPSG: 4326) coordinate reference system.}
}
}
}
\usage{
ch_locations
}
\description{
A dataset containing geographic information about 26 locations in Switzerland. This data includes unique identifiers, names, and geographic coordinates (latitude and longitude) for each location.
}
\keyword{datasets}
