% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowmapblue.R
\name{flowmapblue}
\alias{flowmapblue}
\title{<Add Title>}
\usage{
flowmapblue(
  locations,
  flows,
  mapboxAccessToken = NULL,
  clustering = TRUE,
  animation = FALSE,
  darkMode = FALSE
)
}
\description{
<Add Description>
}
